Fixed bug where deleted entries would cause an infinite loop to occur
The directory entry parser could not properly handle the 0xE5 code and would end up in an infinite loop because the previous name was never cleared.  In the long run, a better way to signal the end of the directoy rather than the name being NULL should be used.  By for now it doesn't break and works properly.